requirements.txt
streamlit>=1.24.1
ultralytics>=8.0.98
opencv-python>=4.7.0.72
Pillow==10.1.0
streamlit_lottie
torch